Update file copying and error handling in AuthentiFile
The change has two purposes. Firstly, the 'CopyToOutputDirectory' setting for 'fileSigs.json' was updated to 'Always' from 'PreserveNewest' in the project file to ensure the file is always copied on build. Secondly, error handling was added to handle exceptions when attempting to read or deserialize 'fileSigs.json' in `AuthentiFile.cs`. If an error occurs, it will now be logged and the program will throw an error. Additionally, a count for masqueraded files was included to provide feedback if no masqueraded files were detected in a given directory.